test(tasks): split routes.test.ts along source seams
The 1,616-line godfile becomes three seam-aligned files plus one shared
harness (SPEC ceiling: no test file over ~1,200 lines):

- routes-rest.test.ts (943) — lib/routes.ts (REST endpoints + registration)
- exec-tools.test.ts (631) — lib/exec-tools.ts (bakin_exec_tasks_*)
- search-doc.test.ts (190) — search index side effects + GET /search
- helpers/tasks-routes-harness.ts — completion-gate fake, seedable run
  state, task-store/task-service mock factories; mock.module wiring stays
  per-file (FW2 B7 precedent)

Pure relocation: 120 pass / 249 expect() calls before and after. Every
split file keeps the FW1.8 dual content-dir mock (both specifiers) and
the dual execution-ledger/task-store mocks.
